requirements.txt
Flask==1.1.2
PyMySQL==0.9.3
SQLAlchemy==1.3.13
Flask-SQLAlchemy==2.4.1
Pillow==7.1.2
gunicorn==20.0.4

